Separate INPUTMASKS to a separate file
